Formatting update to the yokogawa class

+++ instrument.py
@@ -125,22 +125,22 @@ class yokogawa(object):
             return rc
         #_____________________________________________________________________________
         def set_output_state(self,value): 
-            cmd = "OUTP:STAT {0:d}".format( float(value) ) 
+            cmd = "OUTP:STAT {0:d}".format( int(value) ) 
             rc  = self.write(cmd) 
             return rc 
         #_____________________________________________________________________________
         def set_clock_time(self,hour,minute,second): 
-            fhr  = float(hour) 
-            fmin = float(minute) 
-            fsec = float(second)  
+            fhr  = int(hour) 
+            fmin = int(minute) 
+            fsec = int(second)  
             cmd  = ':SYST:CLOC:TIME "{0:02d}:{1:02d}:{2:02d}"'.format(fhr,fmin,fsec)
             rc   = self.write(cmd) 
             return rc 
         #_____________________________________________________________________________
         def set_clock_date(self,month,day,year):
-            fmon = float(month) 
-            fday = float(day) 
-            fyr  = float(year)  
+            fmon = int(month) 
+            fday = int(day) 
+            fyr  = int(year)  
             cmd  = ':SYST:CLOC:DATE "{0:04d}/{1:02d}/{2:02d}"'.format(fyr,fmon,fday)
             rc   = self.write(cmd) 
             return rc
